Fix NorMuon dtype: cast v_mean to float32 for lerp_

+++ records/track_non_record_16mb/2026-03-19_DepthRecurrence_3x3x1024/train_gpt.py
@@ -170,7 +170,7 @@ class Muon(torch.optim.Optimizer):
                     # NorMuon: adaptive column-wise normalization
                     vnorm = g.norm()
                     v_mean = (g * g).mean(dim=-1, keepdim=True)
-                    state["second_momentum"].lerp_(v_mean, 1.0 - beta2)
+                    state["second_momentum"].lerp_(v_mean.float(), 1.0 - beta2)
                     step_size = 1.0 / (state["second_momentum"].sqrt() + 1e-10)
                     g = g * step_size
                     g = g * (vnorm / (g.norm() + 1e-10))
